Refatoração e melhora no menu
Refatorado a parte do desenho do labirinto. Realizado uma melhora no menu, agora mais robusto e sendo o main do programa
