+++ src/Labirinto.h
@@ -9,6 +9,7 @@ typedef struct cell {
     bool isstart;   // eComeço
     bool isdoor;     // ePorta
     bool iskey;     // eChave
+    bool isvisited; // eVisitado
 } celula;
 
 typedef celula** labirinto;
@@ -17,5 +18,3 @@ labirinto Alocar_Labirinto(int linhas, int colunas);
 void Preencher_Labirinto(labirinto* tabuleiro, int linhas, int colunas, FILE *arquivo);
 void Destrutor_Labirinto(labirinto tabuleiro, int linhas);
 void Imprimir_Labirinto(labirinto tabuleiro, int linhas, int colunas);
-void Grafico_Labirinto(char** grafico, int linhas, int colunas);
-void Alocar_Labirinto_Grafico(char** desenho,int linhas, int colunas, labirinto tabuleiro);
